backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Task Queue
celery[redis]>=5.3.0

# Technical Analysis
pandas>=2.1.0
numpy>=1.26.0
ta>=0.10.0
# ta-lib>=0.4.28  # Requires system library installation

# Market Data APIs
yfinance>=0.2.35
polygon-api-client>=1.13.0
alpha-vantage>=2.3.1

# SEC Filings
sec-api>=1.0.17

# Reddit API
praw>=7.7.0

# Sentiment Analysis
transformers>=4.36.0
torch>=2.1.0

# Notifications
aiosmtplib>=3.0.0
twilio>=8.11.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
